Log basic activity counts in redis. Keys:
counts:extractions
counts:keys
counts:tokens
counts:bytes:conversions
counts:bytes:conversions

In the case of logging bytes, for external URLs a HEAD request is done on the remote resource and the Content-Length header is used. This could potentially be off in the case of chunked encodings and how the remote server handles the Content-Length header. In the case of direct uploads the size of the body of the request is used. This could potentially be larger depeneding on the body encoding.