Store timestamped events info in redis using sorted sets and hashes. Includes timestamp, user, type (conversion or extraction for now), file URI, client ip.
New TokenFilter to not only check if the token is valid but also set the X-BD-Username request header and forward that on. This can also be used by down the stream clients (Datawolf, Polyglot, Clowder) to identify the user.

New python test to make sure that deleting a token stops access to resource using the new TokenFilter.